Rework the donations sidebar on the donate page.
Now always shows unsent donations in the sidebar, even when you refresh the entire page. Rendering is separated out in JavaScript to make it easier to add complexity/features to the sidebar. Pretty soon we should just integrate Backbone or something and have real MVC, but this is a step in the right direction.