Only show link to request detail when appropriate
On a fulfillment item, only show the link to the request detail page if
the user is actually the fulfiller. In particular, don't show it for
unfulfilled requests (where the fulfillment is actually a new, unsaved
record), because the volunteer won't be able to see the request until
the actually become the fulfiller.