Improve tests for User#can_request?
I was afraid there was a bug that would prevent users from making a new
request if they had a canceled one. Wrote these tests and found that the
bug did *not* exist. Keeping these tests because they're clearer than
the old ones (and I'm trying to migrate everything to factories anyway).