Change some time intervals from months to days
Time intervals like 1.month in Rails are tricky because not all months
are the same. E.g.:

    now = Time.now.utc
    monthago = now - 1.month
    now - monthago == 1.month  # <-- might not be true

Days are all the same (or at least the discrepancies--leap seconds,
perhaps?--are much smaller and much more rare), so we stick with time
intervals in days instead, like 30.days instead of 1.month.