Update how outstanding donations are displayed
This update starts to eliminate the differences between send_books and
send_money mode. Instead of modes, now every (eligible) donation can be
either sent by the donor, or paid for instead.

Regardless of the user's or donation's mode, the donor is assumed to be
the sender unless and until they pay for the donation.

Now when we display outstanding donations on the homepage, they are
displayed without addresses visible, but with "Send" and "Pay" buttons.
"Send" makes the address visible and also the "Book sent" button.
Alongside this are just the links you need relevant to sending: the
Amazon link, and the flag link.

"Pay" marks the book for payment, kind of like "add to cart". You can
then "pay for selected". This doesn't work yet--that's in a forthcoming
commit. You can also "pay for all" if you don't have anything selected.

There's also a new "outstanding donations" page that just lists the
outstanding donations, in the same format.

On the list of all donations, these controls aren't visible. All
donations are shown with their status. If there's anything outstanding,
it just shows up as "Not sent", and there's a warning/link to the page
where you can take care of them.