Support for paying Orders with Contributions
Redid Order model again. This time removing all stored fields. An Order
just tracks a set of donations to be paid together, and has handy
methods to calculate things and pay them, and that's all.

When a contribution is made in the context of an order, it will now be
associated with the order and will go to pay for that order.

The order detail page is also the Amazon payment return/abandon page,
and handles those cases.